shortened circuits
I made some improvements on the circuit depth
for the OA operator for the fermion and scalar
cases.  I also have verified the matrix element
encoding, and how to use the arccos argument to
get the desired value in the matrix.

I have vaugely tested the staggered phase.  It is producing
minuses on certain elements but I'm not sure if they are the
correct elements.  But I think the staggered phase circuit
is doing what I want it to do.